backend/requirements.txt
Flask==3.0.3
beautifulsoup4==4.12.3
requests==2.32.3
selenium==4.25.0
lxml==5.3.0
html5lib==1.1
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
python-dotenv==1.0.1
fake-useragent==1.5.1
regex==2024.9.11
gunicorn==23.0.0
webdriver-manager==4.0.2
Jinja2==3.1.4
psycopg2-binary==2.9.9
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
authlib==1.3.0
werkzeug==2.3.0
Flask-CORS==4.0.0